major speed up
Former-commit-id: 22a482d0c320e04ed3f6fc4363b3f1f9e7272496

+++ OPAL/br/src/main/scala/org/opalj/fpcf/properties/CallersProperty.scala
@@ -246,7 +246,7 @@ object CallersProperty extends CallersPropertyMetaInformation {
     def toLong(methodId: Int, pc: Int): Long = {
         assert(pc >= 0 && pc < 0xFFFF)
         assert(methodId >= 0 && methodId <= 0x3FFFFF)
-        (methodId.toLong & 0x3FFFFF) | (pc << 22)
+        (methodId.toLong & 0x3FFFFF) | (pc.toLong << 22)
     }
     def toMethodAndPc(pcAndMethod: Long): (Int, Int) = {
         (pcAndMethod.toInt & 0x3FFFFF, (pcAndMethod >> 22).toInt)
